FEAT: Move checkNilArgs in common.go
